requirements.txt
pytesseract==0.3.10
Pillow==10.1.0
biopython==1.81
scikit-learn==1.3.2
numpy==1.26.2